Add method to parse XML booleans
Currently we don't convert boolean objects returned by the server to the
Ruby representation of booleans. This complicates further treatment of
these values, as they have to be treated as strings. To avoid that this
patch introduces a new "parse_bool" method in the base object class
intended to convert the boolean representation used by the server into
the Ruby representation. This method will be used in later changes.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>