Add support for Data Center "local" attribute
In version 3.4 of oVirt Engine the concept of type of a Data Center has
been relaxed, so that multiple types of storage (local, NFS, Fibre
Channel, etc) can be used in the same Data Center. So Data Centers no
longer have a type, but a "local" flag tha indicates if the storage is
only locally accessible or if it may be shared by multiple hosts. This
flag isn't currently available in the Ruby API. This patch adds it,
preserving the existing "storage_type" attribute for backwards
compatibility.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>