Retrieve actual server CA certificate
Currently the tests retrieve the CA certificate from the /ca.crt URL,
and then use it to verify the server certificate in the all the tests.
This will usually work, but it will fail if the user has replaced the
web server certificates with custom ones after the initial installation.
To avoid this issue this patch changes the tests so that they get the
actual CA certificate provided by the server during the SSL handshake.
This will work regardless of what certificate the server is using.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>