Support operating systems
The client now allow to call `operating_systems` to return all
available operating systems. The name then can be passed os value for
`os_type` of vm. Also, I've added support for passing os params
that are compatible with the form we get from vm details:

    :os => { :type => 'rhel_7x86',
             :boot => ['network','hd'],
             :kernel => '..',
             :initrd => '..',
             :cmdline => '.. }

The main reason is to make this option configurable even when going
over fog (that has the :os attribute defined, but not the `:os_type`
and others.