Explicitly use version 3 of the API (#109)
Version 4 of the oVirt engine will support two versions of the API:
version 3, which will be backwards compatible with older versions of the
engine and version 4, which won't be compatible. The default used by the
server will be version 4, but clients can explicitly request a version
of the API using the "Version" header or the "/v3" URL prefix. The
header is the preferred mechanism because it is just ignored by older
versions of the server. This patch modifies the the gem so that it
always send the header explicitly, so that it will work with version 4
of the engine.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit ccec8ac6f26adf6b4cac29a38e0d30f9bc40299e)